docs/examples/zoo/requirements.txt
numpy>=2.0
torch>=1.9.0
nvidia-dali-cuda120>=1.48.0
ffmpeg-python>=0.2.0
pillow
